// src/requests/Request.enum.ts
enum RequestEnum {
  CommonRequest = 'CommonRequest',
  AuthRequest = 'AuthRequest',
  ProjectsRequest = 'ProjectsRequest',
  OrganizationsRequest = 'OrganizationsRequest',
  UsersRequest = 'UsersRequest',
  MarketplaceAccessRequest = 'MarketplaceAccessRequest',
  MastercardRequest = 'MastercardRequest',
  NftRequest = 'NftRequest'
}

export default RequestEnum;